\encoding{latin1}
\name{st-package}
\alias{st-package}
\docType{package}
\title{The st package}

\description{
This package implements the "shrinkage t" statistic 
   described in Opgen-Rhein and Strimmer (2007).  It also offers
   a convenient interface to a number of other regularized t-type 
   statistics often used in high-dimensional case-control studies.
}

\author{Rainer Opgen-Rhein (\url{http://opgen-rhein.de}) and 
        Korbinian Strimmer (\url{http://strimmerlab.org/})}
\references{
  See website: \url{http://strimmerlab.org/software/st/}
}
\keyword{univar}
\seealso{
\code{\link{studentt.stat}, \link{shrinkt.stat}}
}
