\encoding{latin1}
\name{st-package}
\alias{st-package}
\docType{package}
\title{The st package}

\description{
This package implements the "shrinkage t" statistic 
   described in Opgen-Rhein and Strimmer (2007) and a shrinkage estimate
   of the "correlation-adjusted t-score" (cat score) introduced in
   Zuber and Strimmer (2009).  It also offers a convenient interface 
   to a number of other regularized t-statistics commonly 
   employed in high-dimensional case-control studies.    
}

\author{Rainer Opgen-Rhein, Verena Zuber, and 
        Korbinian Strimmer (\url{http://strimmerlab.org/})}
\references{
  See website: \url{http://strimmerlab.org/software/st/}
}
\keyword{univar}
\seealso{
\code{\link{shrinkt.stat}}, 
\code{\link{shrinkcat.stat}}, 
\code{\link{modt.stat}}, 
\code{\link{cst.stat}}, 
\code{\link{lait.stat}}.
}

